+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K83MSATKVG65V1626Q5FV34S"
+        "01K855G3ZJKD15Y1XA9G789QZD"
     )
     r"""The id of the resource"""