+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -136,7 +136,7 @@ DeploymentsModelType = Literal[
     "tts",
     "stt",
     "rerank",
-    "moderations",
+    "moderation",
     "vision",
 ]
 r"""The modality of the model"""