+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-21T15:04:10.057Z")
+    created: Optional[datetime] = parse_datetime("2025-10-22T05:15:28.733Z")
     r"""The date and time the resource was created"""