+++ packages/orq-rc/src/orq_ai_sdk/models/getallpromptsop.py
@@ -58,7 +58,7 @@ GetAllPromptsModelType = Literal[
     "tts",
     "stt",
     "rerank",
-    "moderations",
+    "moderation",
     "vision",
 ]
 r"""The modality of the model"""