+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -38,7 +38,7 @@ GetOnePromptModelType = Literal[
     "tts",
     "stt",
     "rerank",
-    "moderations",
+    "moderation",
     "vision",
 ]
 r"""The modality of the model"""