+++ packages/orq-rc/src/orq_ai_sdk/models/getpromptversionop.py
@@ -67,7 +67,7 @@ GetPromptVersionModelType = Literal[
     "tts",
     "stt",
     "rerank",
-    "moderations",
+    "moderation",
     "vision",
 ]
 r"""The modality of the model"""