+++ packages/orq-rc/src/orq_ai_sdk/models/listpromptversionsop.py
@@ -60,7 +60,7 @@ ListPromptVersionsModelType = Literal[
     "tts",
     "stt",
     "rerank",
-    "moderations",
+    "moderation",
     "vision",
 ]
 r"""The modality of the model"""