+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K83MSATJZS7C2T5H0QKFXHA3"
+        "01K855G3ZJNF8GZZ0XR7R3ZETJ"
     )
     r"""The id of the resource"""