+++ packages/orq-rc/src/orq_ai_sdk/models/updatepromptop.py
@@ -577,7 +577,7 @@ ModelType = Literal[
     "tts",
     "stt",
     "rerank",
-    "moderations",
+    "moderation",
     "vision",
 ]
 r"""The modality of the model"""
@@ -1646,7 +1646,7 @@ UpdatePromptModelType = Literal[
     "tts",
     "stt",
     "rerank",
-    "moderations",
+    "moderation",
     "vision",
 ]
 r"""The modality of the model"""